Split solana_sdk.h
