+++ programs/bpf/c/src/deprecated_loader/deprecated_loader.c
@@ -2,8 +2,9 @@
  * @brief Example C-based BPF program that prints out the parameters
  * passed to it
  */
-#include <solana_sdk.h>
-#include <deserialize_deprecated.h>
+#include <sol_types.h>
+#include <sol_log.h>
+#include <sol_deserialize_deprecated.h>
 
 extern uint64_t entrypoint(const uint8_t *input) {
   SolAccountInfo ka[1];
