+++ programs/bpf/c/src/noop/noop.c
@@ -2,7 +2,8 @@
  * @brief Example C based BPF program that prints out the parameters
  * passed to it
  */
-#include <solana_sdk.h>
+#include <sol_deserialize.h>
+
 
 extern uint64_t entrypoint(const uint8_t *input) {
   SolAccountInfo ka[1];
